% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_game.R
\name{evo_game}
\alias{evo_game}
\title{Simulate a two dimensional plane with agents and food over time and also return meta data}
\usage{
evo_game(
  rowsField = 20,
  columnsField = 20,
  nFood = 50,
  nAgent = 20,
  tick = 20,
  foodGeneratingType = "linear",
  foodGeneratingFactor = 1
)
}
\arguments{
\item{rowsField}{Numeric value > 1 that determines the number of rows the game takes place in.}

\item{columnsField}{Numeric value > 1 that determines the number of columns the game takes place in.}

\item{nFood}{Numeric value that determines the number of food palets that are present at the start of the game.}

\item{nAgent}{Numeric value that determines the number of agents that are present at the start of the game.}

\item{tick}{Numeric value that determines the number of time points that the game runs at most.}

\item{foodGeneratingType}{String value "linear" or "cumulative", which determines if food is added per time point as constant or based on the amount of food that is already present.}

\item{foodGeneratingFactor}{Numeric value that determines how many food is added, depending on foodGeneratingType.}
}
\value{
A list of two data frames with information about individual agents ("Agent_Details") and food quantities ("Food_Details") at each time point, equal to \emph{base_evo_game}. Additionally provides a data frame about meta data ("Meta_Data").
}
\description{
The function \emph{evo_game}, as \emph{base_evo_game} gives information about agents and food per time point and additionally returns meta data for each time point.
}
\details{
For information each data frame column and information about the underlying generating process, see the github README.md file \url{https://github.com/PhilNrbts/Evo_simulation/blob/master/README.md}.
}
\examples{
game <- base_evo_game(rowsField = 4, columnsField = 4, nFood = 2, nAgent = 3, tick = 5, foodGeneratingType = "linear", foodGeneratingFactor = 1)
game$Agent_Details
game$Food_Details
game$Meta_Data
}
\seealso{
\code{\link{base_evo_game}}
}
