% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoGame.R
\docType{package}
\name{EvoGame}
\alias{EvoGame}
\title{This package generates an evolution process for agents}
\description{
This package contains two functions that generate the data about the evolution process. One provides information about agents and food, a second additional meta data information for each time point.
}
\details{
For information each data frame column and information about the underlying generating process, see the github README.md file \url{https://github.com/PhilNrbts/Evo_simulation/blob/master/README.md}.
}
\section{Generate the basic evolution game data}{


\emph{base_evo_game} generates two data frames with information about agents and food for each time point
}

\section{Generate evolution game data and meta data}{


\emph{evo_game} generates information the same information as \emph{base_evo_game} and additionally meta data that summarises information about agents and food for each time point.
}

